feat: embed art for day 228

+++ simtools.py
@@ -216,4 +216,4 @@ def get_embedding(text: str) -> NDArray[np.float32]:
 
 def cosine_similarity(vec1: NDArray[np.float32], vec2: NDArray[np.float32]) -> float:
     """Calculate cosine similarity between two vectors."""
-    return np.dot(vec1, vec2) / (np.linalg.norm(vec1) * np.linalg.norm(vec2))
+    return float(np.dot(vec1, vec2) / (np.linalg.norm(vec1) * np.linalg.norm(vec2)))
